airquality-api/requirements.txt
boto3==1.20.14
botocore==1.23.14
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Flask==2.0.2
flask-lambda-python36==0.1.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
typing-extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
